🔵 reorganize backend (separate feature sets)
