+++ packages/backend/src/domain/contact/spi.ts
@@ -1,0 +1,6 @@
+import { Result } from 'true-myth'
+import { ContactId } from './types'
+
+export type ContactInfoRepository = {
+  add: (email: string, attributes: object) => Promise<Result<ContactId, Error>>
+}
