+++ packages/backend/src/domain/establishment/spi.ts
@@ -1,0 +1,6 @@
+import { Result } from 'true-myth'
+import { Etablissement } from './types'
+
+export type EtablissementRepository = {
+  get: (siret: string) => Promise<Result<Etablissement, Error>>
+}
