+++ packages/backend/src/domain/establishment/types.ts
@@ -93,54 +93,4 @@ export interface Etablissement {
   }>
 }
 
-export interface ContactInfoBodyAttributes {
-  NOM: string
-  PRENOM: string
-  TEL: string
-  SIRET?: string
-  OPT_IN: boolean
-  FORM_NEEDS?: string
-  PROJECT_NEEDS?: string
-  PROJECT_SECTORS?: string[]
-  USER_ROLES?: string
-  USER_GOALS?: string
-  STRUCTURE_SIZE?: string | number
-  PROGRAM_ID?: string
-  ALL_RESPONSES?: string
-}
-
-export interface ContactInfoBody {
-  email: string
-  listIds: number[]
-  attributes: ContactInfoBodyAttributes
-}
-
-export interface ContactId {
-  id: number
-}
-
-class CustomError extends Error {
-  constructor(...args: Array<string | undefined>) {
-    super(...args)
-    this.name = this.constructor.name
-    Error.captureStackTrace(this, CustomError)
-  }
-}
-
-export interface QuestionnaireData {
-  codeNaf?: string
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  [key: string]: any
-}
-
-export interface PublicodesInputData {
-  'entreprise . code NAF'?: string
-  'dispositif . début de validité'?: string
-  'dispositif . fin de validité'?: string
-  'date du jour': string
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  [key: string]: any
-}
-
 export class EstablishmentNotFoundError extends CustomError {}
-export class ServiceNotFoundError extends CustomError {}
