+++ packages/backend/src/domain/program/spi.ts
@@ -1,0 +1,3 @@
+export type CurrentDateService = {
+  get: () => string
+}
