+++ packages/backend/src/domain/services/contact-service.ts
@@ -1,0 +1,14 @@
+import { addBrevoContact } from '@tee/backend/src/infrastructure/brevo-API'
+import { createService } from '../contact/contact-features'
+import { ContactInfoRepository } from '../contact/spi'
+
+/**
+ * Defines how to access external data.
+ * Uses the "Repository" pattern, see README.md
+ */
+const brevoRepository: ContactInfoRepository = {
+  add: addBrevoContact
+}
+const service = createService(brevoRepository)
+
+export const postNewContact = service.postNewContact
