+++ packages/backend/src/domain/services/establishment-service.ts
@@ -1,0 +1,15 @@
+import { createService } from '../establishment/establishment-features'
+import { getEtablissement } from '@tee/backend/src/infrastructure/sirene-API'
+import { EtablissementRepository } from '../establishment/spi'
+
+/**
+ * Defines how to access external data.
+ * Uses the "Repository" pattern, see README.md
+ */
+const etablissementRepository: EtablissementRepository = {
+  get: getEtablissement
+}
+
+const service = createService(etablissementRepository)
+
+export const fetchEtablissement = service.fetchEtablissement
