+++ packages/backend/src/domain/services/program-service.ts
@@ -1,4 +1,6 @@
 import { currentDateService } from '../../infrastructure/current-date'
-import { createService as createFilterService } from '../filter-programs'
+import { createService as createFilterService } from '../program/filter-programs'
+
+export { sortPrograms } from '../program/sort-programs'
 
 export const filterPrograms = createFilterService(currentDateService)
