--- packages/backend/src/domain/spi.ts
@@ -1,14 +1,0 @@
-import { Result } from 'true-myth'
-import { Etablissement, ContactId } from './types'
-
-export type EtablissementRepository = {
-  get: (siret: string) => Promise<Result<Etablissement, Error>>
-}
-
-export type ContactInfoRepository = {
-  add: (email: string, attributes: object) => Promise<Result<ContactId, Error>>
-}
-
-export type CurrentDateService = {
-  get: () => string
-}
