+++ packages/backend/src/infrastructure/current-date.ts
@@ -1,4 +1,4 @@
-import type { CurrentDateService } from '../domain/spi'
+import type { CurrentDateService } from '../domain/program/spi'
 
 export const currentDateService: CurrentDateService = {
   get: () => new Date().toLocaleDateString('fr-FR')
