+++ packages/web/src/stores/programs.ts
@@ -5,9 +5,8 @@ import { computed, ref } from 'vue'
 import { defineStore } from 'pinia'
 
 import type { ProgramData, TrackId, TrackOptions, UsedTrack } from '@/types/index'
-import { filterPrograms as filterWithPublicodes } from '@tee/backend/src/domain/services/filter-service'
-import { sortPrograms } from '@tee/backend/src/domain/sort-programs'
-import type { QuestionnaireData } from '@tee/backend/src/domain/types'
+import { filterPrograms as filterWithPublicodes, sortPrograms } from '@tee/backend/src/domain/services/program-service'
+import type { QuestionnaireData } from '@tee/backend/src/domain/program/types'
 import { TrackHelpValue } from '@/types/index'
 
 export const programsStore = defineStore('programs', () => {
